Added distinction between focused options and selected options.
The below changes improved usability for keyboard navigation but means that the 'BeforeChange' and 'Change' events are now triggered before the 'BeforeClose' event (instead of after).

Changes:
- Added '_focus()' function that applies '.focused' to the currently keyboard-highlighted option;
- The '_focus()' function can optionally trigger the '_select()' function;
- Moved option selecting and label changing from '_close()' to a new '_change()' function;
- The '_change()' function is now triggered by the '_select()' function.

The '.selected' class is only applied on '_close()' to mark the selected option.
A '.focused' class is used on '_select()' to highlight the currently focused option (useful for keyboard navigation).