Fixed lcdsantos/jQuery-Selectric#103
Selectrified elements are now focusable with `<label>` elements. The solution was to add 'focus' events on the original, hidden, `<select>` elements. Upon focus (from the label), the focus is shifted to the hidden `<input>` associated with the Selectric UI.

An alternative solution would be to move the ID from the original `<select>` to the hidden `<input>` field.